+++ src/main/scala/dk/aau/modelardb/Interface.scala
@@ -15,12 +15,16 @@
 package dk.aau.modelardb
 
 import com.sun.net.httpserver.{HttpExchange, HttpHandler, HttpServer}
+import java.net.{ServerSocket, InetSocketAddress}
+
 import dk.aau.modelardb.core.Configuration
 import dk.aau.modelardb.core.utility.Static
 
+import java.io.{BufferedReader, InputStreamReader, PrintWriter}
 import java.nio.file.{Files, Paths}
 import java.util.concurrent.Executor
 import scala.io.Source
+import scala.collection.mutable
 import scala.io.StdIn.readLine
 
 object Interface {
@@ -31,73 +35,73 @@ object Interface {
       return
     }
 
-    this.sql = sql
-    configuration.getString("modelardb.interface") match {
-      case "socket" => socket(configuration.getExecutorService)
-      case "http" => http(configuration.getExecutorService)
-      case "repl" => repl(configuration.getStorage)
-      case path if Files.exists(Paths.get(path)) => file(path)
+    val (interface, port) = this.getInterfaceAndPort(configuration)
+    interface match {
+      case "socket" => socket(configuration.getExecutorService, port, sql)
+      case "http" => http(configuration.getExecutorService, port, sql)
+      case "repl" => repl(configuration.getStorage, sql)
+      case path if Files.exists(Paths.get(path)) => file(path, sql)
       case _ => throw new java.lang.UnsupportedOperationException("unknown value for modelardb.interface in the config file")
     }
   }
 
   /** Private Methods **/
-  private def socket(executor: Executor): Unit = {
+  private def getInterfaceAndPort(configuration: Configuration): (String, Int) = {
+    val interface = configuration.getString("modelardb.interface")
+    val startOfPort = interface .lastIndexOf(':')
+    if (startOfPort == -1) {
+      (interface, 9999)
+    } else {
+      (interface.substring(0, startOfPort), interface.substring(startOfPort + 1).toInt)
+    }
+  }
+
+  private def socket(executor: Executor, port: Int, sql: String => Array[String]): Unit = {
     //Setup
-    val serverSocket = new java.net.ServerSocket(9999)
+    val serverSocket = new ServerSocket(port)
+    Static.info(s"ModelarDB: socket end-point is ready (Port: $port)")
 
     while (true) {
-      Static.info("ModelarDB: socket end-point is ready (Port: 9999)")
       val clientSocket = serverSocket.accept()
       executor.execute(() => {
-        val in = new java.io.BufferedReader(new java.io.InputStreamReader(clientSocket.getInputStream))
-        val out = new java.io.PrintWriter(clientSocket.getOutputStream, true)
-
         //Query
-        Static.info("ModelarDB: connection is ready")
+        val in = new BufferedReader(new InputStreamReader(clientSocket.getInputStream))
+        val out = new PrintWriter(clientSocket.getOutputStream, true)
 
         try {
-          var stop = true
-          while (stop) {
-            val query = in.readLine().trim()
-            if ( ! query.startsWith("--") && query.contains("SELECT")) {
-              execute(query, out.write)
-              out.flush()
-            } else if (query.nonEmpty) {
-              in.close()
-              out.close()
-              clientSocket.close()
-              stop = false //The empty string terminates the connection
-              Static.info("ModelarDB: conection is closed")
-            } else {
-              out.write("only SELECT is supported")
-              out.flush()
-            }
+          while (true) {
+            execute(in.readLine(), sql, out.write)
+            out.flush()
           }
         } catch {
           case _: NullPointerException =>
             //Thrown if the client closes in while ModelarDB waits for input
         }
+
+        out.close()
+        in.close()
+        clientSocket.close()
       })
     }
 
     //Cleanup
     serverSocket.close()
+    Static.info(s"ModelarDB: socket end-point is closed (Port: $port)")
   }
 
-  private def http(executor: Executor): Unit = {
+  private def http(executor: Executor, port: Int, sql: String => Array[String]): Unit = {
     //Setup
-    val server = HttpServer.create(new java.net.InetSocketAddress(9999), 0)
+    val server = HttpServer.create(new InetSocketAddress(port), 0)
 
     //Query
     class QueryHandler extends HttpHandler {
       override def handle(httpExchange: HttpExchange): Unit = {
         val request = httpExchange.getRequestBody
-        val reader = new java.io.BufferedReader(new java.io.InputStreamReader(request))
+        val reader = new BufferedReader(new InputStreamReader(request))
 
         //The query is executed with the result returned as an HTTP response
-        val results = scala.collection.mutable.ArrayBuffer[String]()
-        execute(reader.readLine.trim(), line => results.append(line))
+        val results = mutable.ArrayBuffer[String]()
+        execute(reader.readLine.trim(), sql, line => results.append(line))
         val out = results.mkString("")
         httpExchange.sendResponseHeaders(200, out.length)
         val response = httpExchange.getResponseBody
@@ -110,31 +114,29 @@ object Interface {
     server.createContext("/", new QueryHandler())
     server.setExecutor(executor)
     server.start()
-    Static.info("ModelarDB: HTTP end-point is ready (Port: 9999)")
-    scala.io.StdIn.readLine() //Prevents the method from returning to keep the server running
+    Static.info(s"ModelarDB: HTTP end-point is ready (Port: $port)")
+    readLine() //Prevents the method from returning to keep the server running
 
     //Cleanup
     server.stop(0)
+    Static.info(s"ModelarDB: HTTP end-point is closed (Port: $port)")
   }
 
-  private def repl(storage: String): Unit = {
+  private def repl(storage: String, sql: String => Array[String]): Unit = {
     val prompt = storage.substring(storage.lastIndexOf('/') + 1) + "> "
     do {
       print(prompt)
-      execute(readLine, print)
+      execute(readLine, sql, print)
     } while(true)
   }
 
-  private def file(path: String): Unit = {
+  private def file(path: String, sql: String => Array[String]): Unit = {
     //This method is only called if the file exist
     val st = System.currentTimeMillis()
     Static.info("ModelarDB: executing queries from " + path)
     val source = Source.fromFile(path)
     for (line: String <- source.getLines()) {
-      val q = line.trim()
-      if ( ! (q.isEmpty || q.startsWith("--"))) {
-        execute(q.stripMargin, print)
-      }
+      execute(line, sql, print)
     }
     source.close()
     val et = System.currentTimeMillis() - st
@@ -142,14 +144,27 @@ object Interface {
     Static.info("ModelarDB: finished all queries after " + jst)
   }
 
-  private def execute(query: String, out: String => Unit): Unit = {
+  private def execute(queries: String,
+    sql: String => Array[String], out: String => Unit): Unit = {
+    //Executes each query if it is supported and not commented out
+    for (queriesSplitPart <- queries.split(";")) {
+      val query = queriesSplitPart.trim()
+      if ( ! query.startsWith("--") && query.startsWith("SELECT")) {
+        unsafeExecute(query, sql, out)
+      }
+    }
+  }
+
+  private def unsafeExecute(query: String,
+    sql: String => Array[String], out: String => Unit): Unit = {
+    //Executes a SQL query without performing any checking
     val st = System.currentTimeMillis()
     var result: Array[String] = null
     try {
       val query_rewritten =
         query.replace("COUNT_S(#)", "COUNT_S(tid, start_time, end_time)")
           .replace("#", "tid, start_time, end_time, mtid, model, gaps")
-      result = this.sql(query_rewritten)
+      result = sql(query_rewritten)
     } catch {
       case e: Exception =>
         e.printStackTrace()
@@ -174,7 +189,4 @@ object Interface {
       out(s"""  ]\n}\n""")
     }
   }
-
-  /** Instance Variables **/
-  var sql: String => Array[String] = _
 }